New SecureProtect configuration directive (sidejacking fix)
This is a defense to "sidejacking", the collection of a session cookie by a host on an unsecure network. When SecureProtect is active, the UserDB login process creates a passhash of the encrypted password. This, along with username, login_table, and a "secret" set in the configuration, is used to check subsequent secure accesses to the catalog.